pyproject.toml
[tool.poetry]
name = "systema"
version = "0.0.1a15"
description = ""
authors = ["Rodrigo Eiti Kimura <rodrigoeitikimura@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
typer = {extras = ["all"], version = "^0.9.0"}
sqlmodel = "^0.0.16"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.2.1"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
httpx = "^0.26.0"
nanoid = "^2.0.0"
bcrypt = "^4.1.2"
textual-dev = "^1.5.1"
uvloop = "^0.19.0"
websockets = "^12.0"
httptools = "^0.6.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
commitizen = "^3.14.1"
ruff = "^0.2.1"
pre-commit = "^3.6.1"
behave = "^1.2.6"
watchfiles = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
systema = "systema.cli.main:app"

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "W", "C901"]
extend-select = ["I"]
ignore = []

[tool.commitizen]
version = "0.0.1a15"
tag_format = "v$version"
version_files = [
    "systema/__version__.py",
    "pyproject.toml:version",
]

[tool.pyright]
include = ["."]
exclude = []
ignore = ["./features/steps/*.py"]
